third_party/WebKit/LayoutTests/external/wpt/html/browsers/the-window-object/window-prototype-chain-expected.txt

This is a testharness.js-based test.
PASS window object
FAIL Window.prototypeassert_equals: expected "[object WindowPrototype]" but got "[object Window]"
PASS Global scope polluter
FAIL EventTarget.prototypeassert_equals: expected "[object EventTargetPrototype]" but got "[object EventTarget]"
PASS Object.prototype
Harness: the test ran to completion.
